Add search function for books
Implemented a search bar function to search for book titles. This is achieved by using a LINQ query to the database with an index method inside BooksController. Additionally, the search title is passed as a query string value on the URL using HTTP post and get. With the URL containing the query string value, the search results can be sent to others with the given URL.