Add new field Status
A new field called Status has been added to the books model. It will be a field with two options, available and reserved to show users the booking status of each book. Note that the status field cannot be edited by the user on the app. Books that are newly created onto the database have a default value of available for status.