Add reserve button to books
The reserve button is created adjacent to the edit button for each book showing up in the bookstore list. Currently it only changes the books status from available to reserved, and throws a problem if the user tries to reserve a reserved book. No booking number is given to the user when they make a book reservation yet.